Ensure shfmt indents using 2 spaces
Adhere to the Google Shell Style Guide and the settings that are already in our EditorConfig. Use the available options with MegaLinter and BASH_SHFMT_ARGUMENTS to ensure MegaLinter does not change our preferred style guide.

Also set APPLY_FIXES_MODE to 'pull_request' instead of the default 'commit' to allow the maintainer to review automatic fixes that are applied with MegaLinter.